Break entities into its own files
